% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_status_summary.R
\docType{class}
\name{CalculationStatusSummary}
\alias{CalculationStatusSummary}
\title{CalculationStatusSummary}
\format{An \code{R6Class} generator object}
\description{
CalculationStatusSummary Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{status}}{character [optional]}

\item{\code{points}}{integer [optional]}

\item{\code{requestTime}}{character [optional]}

\item{\code{lastPollTime}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CalculationStatusSummary$new()}}
\item \href{#method-toJSON}{\code{CalculationStatusSummary$toJSON()}}
\item \href{#method-fromJSON}{\code{CalculationStatusSummary$fromJSON()}}
\item \href{#method-toJSONString}{\code{CalculationStatusSummary$toJSONString()}}
\item \href{#method-fromJSONString}{\code{CalculationStatusSummary$fromJSONString()}}
\item \href{#method-clone}{\code{CalculationStatusSummary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalculationStatusSummary$new(
  status = NULL,
  points = NULL,
  requestTime = NULL,
  lastPollTime = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalculationStatusSummary$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalculationStatusSummary$fromJSON(CalculationStatusSummaryJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalculationStatusSummary$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalculationStatusSummary$fromJSONString(CalculationStatusSummaryJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalculationStatusSummary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
