% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_unit_status.R
\docType{class}
\name{CalculationUnitStatus}
\alias{CalculationUnitStatus}
\title{CalculationUnitStatus}
\format{An \code{R6Class} generator object}
\description{
CalculationUnitStatus Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{status}}{character [optional]}

\item{\code{points}}{integer [optional]}

\item{\code{error}}{character [optional]}

\item{\code{result}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CalculationUnitStatus$new()}}
\item \href{#method-toJSON}{\code{CalculationUnitStatus$toJSON()}}
\item \href{#method-fromJSON}{\code{CalculationUnitStatus$fromJSON()}}
\item \href{#method-toJSONString}{\code{CalculationUnitStatus$toJSONString()}}
\item \href{#method-fromJSONString}{\code{CalculationUnitStatus$fromJSONString()}}
\item \href{#method-clone}{\code{CalculationUnitStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalculationUnitStatus$new(
  status = NULL,
  points = NULL,
  error = NULL,
  result = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalculationUnitStatus$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalculationUnitStatus$fromJSON(CalculationUnitStatusJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalculationUnitStatus$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalculationUnitStatus$fromJSONString(CalculationUnitStatusJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalculationUnitStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
