% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_account.R
\docType{class}
\name{ConfigurationAccount}
\alias{ConfigurationAccount}
\title{ConfigurationAccount}
\format{An \code{R6Class} generator object}
\description{
ConfigurationAccount Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{benchmarkCode}}{character [optional]}

\item{\code{benchmarkName}}{character [optional]}

\item{\code{maxEndDate}}{character [optional]}

\item{\code{minStartDate}}{character [optional]}

\item{\code{lockingDate}}{character [optional]}

\item{\code{name}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConfigurationAccount$new()}}
\item \href{#method-toJSON}{\code{ConfigurationAccount$toJSON()}}
\item \href{#method-fromJSON}{\code{ConfigurationAccount$fromJSON()}}
\item \href{#method-toJSONString}{\code{ConfigurationAccount$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ConfigurationAccount$fromJSONString()}}
\item \href{#method-clone}{\code{ConfigurationAccount$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationAccount$new(
  benchmarkCode = NULL,
  benchmarkName = NULL,
  maxEndDate = NULL,
  minStartDate = NULL,
  lockingDate = NULL,
  name = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationAccount$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationAccount$fromJSON(ConfigurationAccountJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationAccount$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationAccount$fromJSONString(ConfigurationAccountJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationAccount$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
