% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_component.R
\docType{class}
\name{PAComponent}
\alias{PAComponent}
\title{PAComponent}
\format{An \code{R6Class} generator object}
\description{
PAComponent Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{accounts}}{list( \link{PAIdentifier} ) [optional]}

\item{\code{benchmarks}}{list( \link{PAIdentifier} ) [optional]}

\item{\code{currencyisocode}}{character [optional]}

\item{\code{dates}}{\link{PADateParameters} [optional]}

\item{\code{snapshot}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{category}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PAComponent$new()}}
\item \href{#method-toJSON}{\code{PAComponent$toJSON()}}
\item \href{#method-fromJSON}{\code{PAComponent$fromJSON()}}
\item \href{#method-toJSONString}{\code{PAComponent$toJSONString()}}
\item \href{#method-fromJSONString}{\code{PAComponent$fromJSONString()}}
\item \href{#method-clone}{\code{PAComponent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PAComponent$new(
  id = NULL,
  accounts = NULL,
  benchmarks = NULL,
  currencyisocode = NULL,
  dates = NULL,
  snapshot = NULL,
  name = NULL,
  category = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PAComponent$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PAComponent$fromJSON(PAComponentJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PAComponent$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PAComponent$fromJSONString(PAComponentJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PAComponent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
