% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spar_calculation_parameters.R
\docType{class}
\name{SPARCalculationParameters}
\alias{SPARCalculationParameters}
\title{SPARCalculationParameters}
\format{An \code{R6Class} generator object}
\description{
SPARCalculationParameters Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{componentid}}{character}

\item{\code{accounts}}{list( \link{SPARIdentifier} ) [optional]}

\item{\code{benchmark}}{\link{SPARIdentifier} [optional]}

\item{\code{dates}}{\link{SPARDateParameters} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SPARCalculationParameters$new()}}
\item \href{#method-toJSON}{\code{SPARCalculationParameters$toJSON()}}
\item \href{#method-fromJSON}{\code{SPARCalculationParameters$fromJSON()}}
\item \href{#method-toJSONString}{\code{SPARCalculationParameters$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SPARCalculationParameters$fromJSONString()}}
\item \href{#method-clone}{\code{SPARCalculationParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPARCalculationParameters$new(
  componentid,
  accounts = NULL,
  benchmark = NULL,
  dates = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPARCalculationParameters$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPARCalculationParameters$fromJSON(SPARCalculationParametersJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPARCalculationParameters$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPARCalculationParameters$fromJSONString(SPARCalculationParametersJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPARCalculationParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
