% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_calculation_parameters.R
\docType{class}
\name{VaultCalculationParameters}
\alias{VaultCalculationParameters}
\title{VaultCalculationParameters}
\format{An \code{R6Class} generator object}
\description{
VaultCalculationParameters Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{componentid}}{character}

\item{\code{account}}{\link{VaultIdentifier}}

\item{\code{dates}}{\link{VaultDateParameters} [optional]}

\item{\code{configid}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VaultCalculationParameters$new()}}
\item \href{#method-toJSON}{\code{VaultCalculationParameters$toJSON()}}
\item \href{#method-fromJSON}{\code{VaultCalculationParameters$fromJSON()}}
\item \href{#method-toJSONString}{\code{VaultCalculationParameters$toJSONString()}}
\item \href{#method-fromJSONString}{\code{VaultCalculationParameters$fromJSONString()}}
\item \href{#method-clone}{\code{VaultCalculationParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultCalculationParameters$new(
  componentid,
  account,
  configid,
  dates = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultCalculationParameters$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultCalculationParameters$fromJSON(VaultCalculationParametersJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultCalculationParameters$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultCalculationParameters$fromJSONString(VaultCalculationParametersJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultCalculationParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
