% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_component.R
\docType{class}
\name{VaultComponent}
\alias{VaultComponent}
\title{VaultComponent}
\format{An \code{R6Class} generator object}
\description{
VaultComponent Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{account}}{\link{VaultIdentifier} [optional]}

\item{\code{benchmark}}{\link{VaultIdentifier} [optional]}

\item{\code{currencyisocode}}{character [optional]}

\item{\code{dates}}{\link{VaultDateParameters} [optional]}

\item{\code{snapshot}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{category}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VaultComponent$new()}}
\item \href{#method-toJSON}{\code{VaultComponent$toJSON()}}
\item \href{#method-fromJSON}{\code{VaultComponent$fromJSON()}}
\item \href{#method-toJSONString}{\code{VaultComponent$toJSONString()}}
\item \href{#method-fromJSONString}{\code{VaultComponent$fromJSONString()}}
\item \href{#method-clone}{\code{VaultComponent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultComponent$new(
  id = NULL,
  account = NULL,
  benchmark = NULL,
  currencyisocode = NULL,
  dates = NULL,
  snapshot = NULL,
  name = NULL,
  category = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultComponent$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultComponent$fromJSON(VaultComponentJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultComponent$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultComponent$fromJSONString(VaultComponentJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultComponent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
