% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_date_parameters.R
\docType{class}
\name{VaultDateParameters}
\alias{VaultDateParameters}
\title{VaultDateParameters}
\format{An \code{R6Class} generator object}
\description{
VaultDateParameters Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{startdate}}{character [optional]}

\item{\code{enddate}}{character}

\item{\code{frequency}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VaultDateParameters$new()}}
\item \href{#method-toJSON}{\code{VaultDateParameters$toJSON()}}
\item \href{#method-fromJSON}{\code{VaultDateParameters$fromJSON()}}
\item \href{#method-toJSONString}{\code{VaultDateParameters$toJSONString()}}
\item \href{#method-fromJSONString}{\code{VaultDateParameters$fromJSONString()}}
\item \href{#method-clone}{\code{VaultDateParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultDateParameters$new(enddate, frequency, startdate = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultDateParameters$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultDateParameters$fromJSON(VaultDateParametersJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultDateParameters$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultDateParameters$fromJSONString(VaultDateParametersJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultDateParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
